# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Halls Head Cricket Club Stats API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stats:
    get:
      operationId: listStats
      tags: [stats]
      summary: List per-grade stats records
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: grade
          in: query
          schema:
            type: string
        - name: playerId
          in: query
          schema:
            type: integer
        - name: season
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [games, runs, wickets, batAvg, bowlAvg, catches, name, season]
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of stat records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatListResponse"
    post:
      operationId: createStat
      tags: [stats]
      summary: Create a stat record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stat"
  /stats/{id}:
    get:
      operationId: getStat
      tags: [stats]
      summary: Get a stat record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stat record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stat"
        "404":
          description: Not found
    patch:
      operationId: updateStat
      tags: [stats]
      summary: Update a stat record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatUpdate"
      responses:
        "200":
          description: Updated stat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stat"
        "404":
          description: Not found
    delete:
      operationId: deleteStat
      tags: [stats]
      summary: Delete a stat record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Stat:
      type: object
      required: [id, playerId, surname, givenName, grade]
      properties:
        id:
          type: integer
        playerId:
          type: integer
        surname:
          type: string
        givenName:
          type: string
        grade:
          type: string
        season:
          type: ["integer", "null"]
        games:
          type: ["integer", "null"]
        innings:
          type: ["integer", "null"]
        notOuts:
          type: ["integer", "null"]
        runs:
          type: ["integer", "null"]
        batAvg:
          type: ["number", "null"]
        highScore:
          type: ["string", "null"]
        fifties:
          type: ["integer", "null"]
        hundreds:
          type: ["integer", "null"]
        wickets:
          type: ["integer", "null"]
        runsConceded:
          type: ["integer", "null"]
        bowlAvg:
          type: ["number", "null"]
        bestBowling:
          type: ["string", "null"]
        fiveWickets:
          type: ["integer", "null"]
        catches:
          type: ["integer", "null"]
        stumpings:
          type: ["integer", "null"]
        runOuts:
          type: ["integer", "null"]
    StatInput:
      type: object
      required: [playerId, grade]
      properties:
        playerId:
          type: integer
        grade:
          type: string
        season:
          type: integer
        games:
          type: integer
        innings:
          type: integer
        notOuts:
          type: integer
        runs:
          type: integer
        batAvg:
          type: number
        highScore:
          type: string
        fifties:
          type: integer
        hundreds:
          type: integer
        wickets:
          type: integer
        runsConceded:
          type: integer
        bowlAvg:
          type: number
        bestBowling:
          type: string
        fiveWickets:
          type: integer
        catches:
          type: integer
        stumpings:
          type: integer
        runOuts:
          type: integer
    StatUpdate:
      type: object
      properties:
        grade:
          type: string
        season:
          type: integer
        games:
          type: integer
        innings:
          type: integer
        notOuts:
          type: integer
        runs:
          type: integer
        batAvg:
          type: number
        highScore:
          type: string
        fifties:
          type: integer
        hundreds:
          type: integer
        wickets:
          type: integer
        runsConceded:
          type: integer
        bowlAvg:
          type: number
        bestBowling:
          type: string
        fiveWickets:
          type: integer
        catches:
          type: integer
        stumpings:
          type: integer
        runOuts:
          type: integer
    StatListResponse:
      type: object
      required: [stats, total, page, limit]
      properties:
        stats:
          type: array
          items:
            $ref: "#/components/schemas/Stat"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...